+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## Unreleased
+### Added
+- check-elb-health-sdk.rb: add option for warning instead of critical when unhealthy instances are found
 
 ## [2.0.1] - 2015-11-03
 ### Changed